backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Payments
stripe==8.0.0

# Scheduling
apscheduler==3.10.4

# Google Sheets
gspread==6.0.2
gspread-formatting==1.1.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# AI (optional)
anthropic==0.18.1
